backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pymysql>=1.1.0
cryptography>=41.0.7
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.1
python-multipart>=0.0.6
httpx>=0.25.2
pycryptodome>=3.19.0
pydantic>=2.9.0
pydantic[email]>=2.9.0